Initital Commit
On this first commit, I have the beginnings of a personal website started. So far I have a nav bar with working links, media queries for accounting for differently sized devices, and incorporation of jquery for forcing the navbar icon to collapse when a link is clicked (this was especially difficult!!!).