Functionality for clicking out of the expanded navbar
Added functionality for a user being able to click out of the expanded navbar in order to get back to the page, similar to how mobile websites work. Also added functionality for making all of the contents on the current page disabled when the navbar is expanded so that when the user clicks out of the navbar, they don't accidentally click any buttons they weren't planning on.